Calculate stats online
Do calculations are results are presented instead of caching them for
later processing. This involves wrapping things in a readline structure.

There are also several refactorings here to simplify the structure.

If the spawn emits an error, the promise rejection will pass that on
directly.

The string-to-Reader patch in the last three test does not work on Node
versions less than 0.11.14. This shouldn't make a difference in
production use and alternatives are a bit gnarly so I've exempted them.